+++ src/components/mat-icon/mat-icon-interface.tsx
@@ -1,0 +1,1 @@
+export type MaterialIconFamily = 'filled' | 'outlined' | 'twotone' | 'rounded' | 'sharp';